Handling B32Muli as a 128bit instruction (#7)
This PR handles B32Muli as a complex instruction, so that we can run
fibonacci with the instruction `B32_MULI @8, @4, #-1G` .

It introduces a test to ensure Fibonacci is parsed and runs smoothly.
It also introduces a quick temporary fix to the frame allocator so that
we can test fibonacci and Collatz. Only the balancing of the state channel is failing in the Collatz test.