Handle CALL procedures and use Allocator (#27)
This PR handles the move operations within a CALL procedure together,
and values are set once all the required information is provided.
For this, it refactors the PROM such that each instruction has the
following structure:
```
struct InterpreterInstruction {
    instruction: [BF16, 4],
    field_pc: BF32,
    is_call_procedure: bool,
}
```
Moreover, it uses the allocator to allocate frames and set the frame
pointer.
It also removes the arguments from LabelsFrameSizes, as they are no
longer useful.

I added `moves_to_set` -- a vector of MVInfo (containing all the info we
have about a certain move) which stores all the moves contained in the
CALL procedure. Then, in the call itself, we decide whether to
immediately generate a move event (when the we already have access to
the value to store), or to add all the necessary info to
`interpreter.vrom.to_set`. `to_set` keeps track of the push events that
need to be set once the source value is set.