Add support for signed integer operations (#56)
This PR changes MUL, MULI, ADD and ADDI so they support signed values,
and adds MULSU and MULU.
Moreover, to abide by the specs, all MUL operations now return 64 bits
(which means the code for Collatz needed to be slightly adapted to take
that into account).
It also adds a unit test for all multiplication opcodes.