Make vrom channel a lookup table and add B32_MUL (#81)
The `vrom` channel is now a lookup channel with a multiplicity field,
meaning that a push to the `vrom` channel can be pulled multiple times.
We still need to convert the `prom` channel to a lookup channel.
Additionally, support for b32_mul has been added in this PR primarily to
test this feature.