Add light-weight ASM frame tooling to make reading VROM easier (#87)
There is going to be a following PR once this is merged that adds logic
to run all of the ASM tests in #67. However, before I start on that, I
think it's a good idea to get this reviewed since the that PR is going
to depend on this.

Essentially this just adds a way to refer to expected VROM values in
tests relative to a allocated frame instead of doing the address
calculations by hand of where each frame value should be.

I also converted `fibonacci.rs` over to use frames.