Add arithmetization for binary ops (#112)
Implements:
- [x] AND
- [x] XOR
- [x] OR
- [x] ORI
- [ ] B32_MULI

It contains a workaround for ORI to accomodate for the projection of
`imm` from 16b to 32b, with a TODO similar to the #105 PR, and I haven't
grouped similar instructions yet because of this, will do it in a
separate PR later.

We can also revamp the `simple` test file to remove a bunch of
boilerplate code later.