Fix the ASM test running using totally incorrect initial frame values (#143)
So... Kind of an interesting accident that the majority of the tests
still managed to pass with this bug, but the test runner was always
converting initial frame values (besides `0` & `1`, which are always set
to `0`) to fields. This PR fixes that issue and all ASM tests except
`bezout` now pass.