Bump nightly (#175)
Toolchain hadn't been updated since the start (`binius`' one has).
Figured we could as well set it to latest to be fine for a while (if we
want to keep a pinned one for reproducibility and consistency).

Allows to simplify the `Table` trait by removing the explicit upcasting
method as pointed out by Sergei on slack.

Also seems `clippy:uninlined_format_args` lint moved to warnings now.