Deterministic assembly parsing support (#186)
This PR augments the parser with the support of "prover view"
instructions (those that come with a `!` bang at the end of their opcode
and are intended for the emulator, not the prover). The boolean field
`prover_only` has been added to the `InstructionWithLabel` clauses where
applicable (all instructions besides conditionals, jumps/calls and
allocations).