Remove nondeterministic move logic (#196)
This PR aims at removing the special MOVE logice, which was made
unnecessary thanks to the `FP` instruction.

Note that for `TAILI`, we can no longer use a MOVE instruction to
directly set the return value *before* the `TAILI` call. Instead, we
compute the absolute address of the value we need to set (using `FP` for
example), and send this pointer to the next frame instead. The called
function can then directly set the value within its body. This new logic
is used, for example, in `fibonacci.asm`.