Change bumpalo to binius HostBumpAllocator (#203)
The latest binius master started using the internal
binius::HostBumpAllocator, so I updated this to match. It uses a
somewhat magic size, but it passes the Fibonacci test with 100,000
iterations, so I hope it should work well.
Also changed OptimalUnderlier128b to OptimalUnderlier to get better
performance with AVX2/AVX512.